Complete Writing and Testing of Aggregation

+++ bm3d.h
@@ -41,7 +41,7 @@ private:
     float* d_numerator;    //Numerator used for aggregation
     uint* d_denominator;  //Denminator used for aggregation
     // cuComplex* d_transformed_stacks;    //3D groups of a batch
-    uint* d_weight;                   //Weights for aggregation
+    float* d_weight;                   //Weights for aggregation
     float* d_wien_coef;             //Only for two step denoising, contains wiener coefficients
     float* d_kaiser_window;         //Kaiser window used for aggregation
     int total_ref_patches;
